requirements.txt
Django==4.2.6
django_environ==0.11.2
djangorestframework==3.14.0
googlemaps==4.10.0
Requests==2.31.0
